Fixing a formatting problem
According to the mock up the Skills portion of the resume should be in column format and not row, but in this file the HTMLskillsStart has the 'ul' element listed as class flex-box which according to the CSS is flex - row. I decided to just remove the class from the 'ul' as opposed to writing a new class.